Moved BCode to dasm.h
